Fix bit_* functions taking int instead of uint64_t (#36)

+++ kernel/types/bit/bit.cpp
@@ -1,17 +1,18 @@
+#include <stdint.h>
 
-void bit_set(unsigned char* bitmap, const int page) {
+void bit_set(unsigned char* bitmap, const uint64_t page) {
     bitmap[page / 8] |= (1 << (page % 8));
 }
 
-void bit_clear(unsigned char* bitmap, const int page) {
+void bit_clear(unsigned char* bitmap, const uint64_t page) {
     bitmap[page / 8] &= ~(1 << (page % 8));
 }
 
-void bit_write(unsigned char* bitmap, const int page, const bool value) {
+void bit_write(unsigned char* bitmap, const uint64_t page, const bool value) {
     if (value) bit_set(bitmap, page);
     else bit_clear(bitmap, page);
 }
 
-bool bit_read(unsigned char* bitmap, const int page) {
+bool bit_read(unsigned char* bitmap, const uint64_t page) {
     return (bitmap[page / 8] >> (page % 8)) & 1;
 }
